Crate timers

A generic timer scheduler module that can be integrated into a crossbeam based event loop or used to launch a background timer thread.

Structs

ScheduledEvent 🔒

TimerEventRequest

Requests a TimerEvent-Message be sent after the given duration.

TimerId

TimerScheduler

A queue of TimerEventRequests that are stored in order of next-to-fire.

Type Aliases

BoxedTimerCallback

A callback to pass to the TimerScheduler to be called when the timer is dispatched.